+++ engine/lib/quoll/asset/AssetRegistry.h
@@ -1,5 +1,6 @@
 #pragma once
 
+#include "quoll/core/Result.h"
 #include "quoll/animation/AnimationAsset.h"
 #include "quoll/animation/AnimatorAsset.h"
 #include "quoll/audio/AudioAsset.h"
@@ -16,7 +17,6 @@
 #include "AssetHandle.h"
 #include "AssetMap.h"
 #include "AssetType.h"
-#include "Result.h"
 
 namespace quoll {
