+++ engine/lib/quoll/core/Result.h
