+++ engine/lib/quoll/io/EntitySerializer.h
@@ -1,6 +1,6 @@
 #pragma
 
-#include "quoll/asset/Result.h"
+#include "quoll/core/Result.h"
 #include "quoll/entity/Entity.h"
 #include "quoll/yaml/Yaml.h"
