+++ engine/lib/quoll/text/MsdfLoader.h
@@ -1,7 +1,7 @@
 #pragma once
 
+#include "quoll/core/Result.h"
 #include "quoll/asset/AssetData.h"
-#include "quoll/asset/Result.h"
 #include "FontAsset.h"
 
 namespace quoll {
